Fixed a bug related to renaming.
In the previous version, when renaming the identifiers, the symbol table
created in the semantic check is used to associate the identifiers with
the declarations. This approach, however, is wrong. Consider the
following example:

int a = 1;
{
  int b = a;
  int a = 2;
}

The "a" in the inner scope will be associated with the latter
declaration instead of the first one, which leads to a bug. In this
version, the associating process is done during the semantic check.