Rewrite some front-end code.
Extract the renaming of identifier part out of SemanticChecker. It is
now moved to the preprocessing part of IRBuilder [cf. preprocessor.h].
Further analysis of the AST will also locate here.

P.S: The old implementation of renaming has a bug. Generate the IR of
testcase_594 and read the first few lines of the main function.